separated the content from the students template comp
